move combined tests to Issues namesapce
