--- rules-tests/Renaming/Rector/FuncCall/RenameFunctionWithRemoveDeadStmtRector/Fixture/fixture.php.inc
@@ -1,27 +1,0 @@
-<?php
-
-namespace Rector\Tests\Renaming\Rector\FuncCall\RenameFunctionWithRemoveDeadStmtRector\Fixture;
-
-class Fixture
-{
-    public function someMethod()
-    {
-        preg_replace('//', 'foo', 'bar');
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\Renaming\Rector\FuncCall\RenameFunctionWithRemoveDeadStmtRector\Fixture;
-
-class Fixture
-{
-    public function someMethod()
-    {
-        \Safe\preg_replace()('//', 'foo', 'bar');
-    }
-}
-
-?>
