+++ tests/Issues/Issue6420/Fixture/fixture.php.inc
@@ -1,0 +1,26 @@
+<?php
+
+namespace Rector\Core\Tests\Issues\Issue6420\Fixture;
+
+class Fixture
+{
+    public function someMethod()
+    {
+        preg_replace('//', 'foo', 'bar');
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Core\Tests\Issues\Issue6420\Fixture;
+
+class Fixture
+{
+    public function someMethod()
+    {
+    }
+}
+
+?>
