+++ tests/Issues/Issue6420/config/configured_rule.php
